+++ NEWS.md
@@ -1,6 +1,6 @@
 # usethis (development version)
 
-* `use_pkgdown()` (and therefore `use_pkgdown_github_pages()`)  now alerts user if they are not on the default branch. (#2065, @mhausinger))
+* `use_pkgdown_github_pages()` now challenges the user when they are not on the default branch, because setting up the site on a non-default branch leads to an empty deployed site (#2065, @mhausinger).
 * `use_snapshot()` is a new function to open a testthat snapshot file corresponding to a given test/R file (#2156, @jonthegeek).
 
 # usethis 3.2.2